refactor: isolate grpc typed rpc adapters
