+++ lmcache/v1/multiprocess/transport/grpc_impl/__init__.py
@@ -1,7 +1,8 @@
 # SPDX-License-Identifier: Apache-2.0
-"""gRPC-generated stubs for the LMCache mp-mode message queue.
+"""gRPC protocol support for the LMCache mp-mode message queue.
 
-Only the generated proto module (``_proto_gen``) is exported now.  The
-old ``GrpcClientTransport`` / ``GrpcServerTransport`` classes are gone;
-``lmcache.v1.multiprocess.mq`` talks to the generated stubs directly.
+Generated stubs live in ``_proto_gen`` and explicit Python/protobuf adapters
+live in ``typed_rpc``. The old ``GrpcClientTransport`` /
+``GrpcServerTransport`` classes are gone; ``lmcache.v1.multiprocess.mq``
+contains only client/server transport control flow.
 """
